favor maps over lists due to helm list merge limitations
Originally, I went with an approach that uses lists of objects.
However, after experiencing some pain with the helm support for combining / overriding lists,
I reverted back to a map-based approach where the key-names have more value.
This is more consistent with the approach we use in fhir-server-config anyway, and so
I went ahead and made them exactly alike so that we can just rely on helm's toJson instead of
writing our own templates for this.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>

bump the chart dependency version

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>