move more keycloak defaults into values.yaml
previously I left the keycloak admin username/password to my overrides file,
but because we need those in the configure-keycloak job they aren't really optional.

additionally, I was using KEYCLOAK_USER and KEYCLOAK_PASSWORD env vars, but
this is bad because KEYCLOAK_PASSWORD ends up in clear text in the keycloak statefulset...
now we populate KEYCLOAK_USER_FILE and KEYCLOAK_PASSWORD_FILE from the same secret
that we're using for the configure-keycloak job.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>